requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
asyncpg==0.29.0

tortoise-orm==0.20.0
asyncpg==0.29.0
aerich==0.7.2

torch==2.5.1
torchvision
torchaudio
transformers==4.36.2
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1

prometheus-client==0.19.0
python-json-logger==2.0.7
psutil==5.9.7

slowapi==0.1.8
python-dotenv==1.0.0
bcrypt==4.1.2

PyYAML==6.0.1
python-dateutil==2.8.2
aiohttp==3.9.1
requests==2.31.0
tqdm==4.66.1

pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
